feat: Implement stamina and injury systems
- Added stamina mechanics tied to the Állóképesség attribute, including fatigue states with progressive combat penalties.
- Integrated stamina costs for attacks, blocks, parries, and dodges.
- Introduced an injury condition system with four tiers (Healthy, Light, Serious, Critical) based on FP and EP thresholds, affecting combat stats.
- Enhanced attack resolution to account for stamina and injury conditions, including unconscious mechanics where units at 0 stamina cannot act.
- Updated combat messages to display hit zones, SFÉ absorption, rolled damage, and stamina costs for defensive actions.
- Improved HUD and unit info popups to reflect stamina and injury conditions accurately.
- Adjusted documentation to reflect new systems and their mechanics.

+++ MAGUS_pygame/domain/mechanics/damage.py
@@ -82,9 +82,9 @@ def calculate_final_damage(
     attr_bonus = _calculate_attribute_bonus(attacker, weapon)
     modified = base_damage + attr_bonus
 
-    # Apply multiplier (e.g., charge)
+    # Apply multiplier (e.g., charge, critical)
     charge_mult = max(1, ctx.charge_multiplier)
-    multiplied = modified * charge_mult
+    multiplied = int(modified * charge_mult)
 
     # Armor absorption (flat)
     absorbed = min(multiplied, max(0, ctx.armor_absorption))
